Thoroughly changed attributes classes

+++ Assets/Scripts/Collectables/Shield.cs
@@ -4,7 +4,7 @@ public class Shield : TimeBonus {
 
     public override void Collect()
     {
-        BonusManager.Instance.IncreaseShieldProtection(bonusValue, bonusTime);
+        BonusManager.Instance.IncreaseShieldProtection((int)bonusValue, bonusTime);
         Destroy(gameObject);
     }
 }
